refactor: update order-related components and restructure imports for improved organization
